fix: Address PR #155 code review comments
- Add PID gain and integral scaling enum constants to eliminate magic numbers
- Replace size_t cast with uint32_t in diagnostics for consistency
- Fix payload buffer size in comm_manager (256 → 418 bytes)
- Fix undefined variable references in test_rx_nanopb.c
- Correct TEST_ASSERT_FLOAT_WITHIN argument count in tests

All changes comply with CLAUDE.md coding standards and NASA Power of 10 rules.

+++ star-rx72n-firmware/src/diagnostics.c
@@ -114,7 +114,7 @@ void diagnostics_log_fault(const char* tag, const char* message, int32_t error_c
                                (unsigned long)error_code);
 
   /* Ensure buffer wasn't truncated */
-  if (written > 0 && (size_t)written < sizeof(log_buffer)) {
+  if (written > 0 && (uint32_t)written < sizeof(log_buffer)) {
     rx_log_error(tag, log_buffer);
   } else {
     /* Fallback if buffer too small */
